API Reference

AbstractTree Functions

Defining Trees

JuDGE.narytree — Function

narytree(depth::Int64, degree::Int64)

Given the depth and degree, this function returns an N-ary tree. Note that a depth of 0 return a single Leaf node (which is also the root node of the tree).

Required Arguments

depth is the maximum number of arcs from the root node any Leaf node

degree is the number of children of all nodes, other than the Leaf nodes

Example

tree = narytree(2,2)

JuDGE.tree_from_file — Function

tree_from_file(filename::String)

Construct tree from a file, each line in the file is of the form B,A,... representing an arc in the tree, from node "A" to node "B". The total number of columns is arbitrary. The first row of the file should be n,p,... these column headers are converted into symbols used to index the data. Each column itself converted into a dictionary, indexed by the node.

Required Arguments

string is the full path of the file containing the tree

Example

tree, data = tree_from_file(joinpath(@__DIR__,"tree.csv"))

JuDGE.tree_from_leaves — Function

tree_from_leaves(leafnodes::Array{Array{Int64,1},1}, probs::Array{Float64,1})

Construct tree from Array of leaf nodes, and (optionally) the corresponding probabilities

Required Arguments

leafnodes is an array of arrays defining the set of leaf nodes

Optional Arguments

probs is an array of probabilities for the leaf nodes

Example

(tree,prob) = tree_from_leaves([[1,1,1],[1,1,2],[1,2,1],[1,2,2]],[0.25,0.25,0.25,0.25])
tree = tree_from_leaves([[1,1,1],[1,1,2],[1,2,1],[1,2,2]])

JuDGE.print_tree — Function

print_tree(some_tree::AbstractTree, data::T where T <: Dict)

Given some_tree, this function prints a simple representation of the tree to the REPL.

Required Arguments

some_tree is the tree we wish to visualise

Optional Arguments

data is a dictionary indexed by the nodes of some_tree

Nodes of Trees

Base.collect — Function

collect(tree::Tree;order=:depth)

Given tree, this function returns an array of corresponding nodes. By default this will be in a depth-first order.

Required Arguments

tree is the tree from which we wish to collect the nodes

Optional Arguments

order can be set to :depth or :breadth to specify the order that the nodes are listed in the array.

Examples

nodes = collect(tree) #gets an array of nodes from tree in depth-first order
nodes = collect(tree,order=:breadth) #gets an array of nodes from tree in breadth-first order

JuDGE.get_leafnodes — Function

get_leafnodes(tree::AbstractTree)

Given tree, this function returns an array of corresponding Leaf nodes.

Required Arguments

tree is the tree from which we wish to collect leaf nodes

Example

leafnodes = JuDGE.get_leafnodes(tree) #define a function that returns the parent of each node in the tree

JuDGE.get_node — Function

get_node(tree::AbstractTree, indices::Array{Int64,1})

Given a tree, and an array of indices, this function returns the corresponding node in the tree.

Required Arguments

tree is the tree from which we are finding the node

indicies is an array of integer indices identifying a node within tree.

Examples

node = get_node(tree,[1]) #get the root node
node = get_node(tree,[1,1]) #get the first child of the root node
node = get_node(tree,[1,2]) #get the second child of the root node

Tree Probabilities

JuDGE.convert_probabilities — Function

convert_probabilities(tree::AbstractTree, probabilities::Dict{AbstractTree,Float64})

Given a dictionary of conditional probabilities for each node in tree, this function returns a dictionary that maps each node of tree to the corresponding unconditional probability.

Required Arguments

tree is the tree that the probabilities pertain to

probabilities is a dictionary of condition probabilities for each node in tree

Example

probs = JuDGE.convert_probabilities(tree,probabilities)

JuDGE.ConditionallyUniformProbabilities — Function

ConditionallyUniformProbabilities(tree::AbstractTree)

Given a tree, this function returns a dictionary which maps nodes of the tree to probabilities, given that there are conditionally uniform probabilities over the children of any node.

Required Arguments

tree is the tree for which the probability distribution will be generated

Example

probs = ConditionallyUniformProbabilities(tree)

JuDGE.UniformLeafProbabilities — Function

UniformLeafProbabilities(tree::AbstractTree)

Given a tree, this function returns a dictionary which maps nodes of the tree to probabilities, given that there is a uniform distribution over the leaf nodes

Required Arguments

tree is the tree for which the probability distribution will be generated

Example

probs = UniformLeafProbabilities(tree)

Other Tree functions

JuDGE.depth — Function

depth(tree::AbstractTree)

Given tree, this function returns a function that takes a node of tree and returns that node's depth. The root node has a depth of 0.

Required Arguments

tree is the tree that the depth function will correspond to.

Example

depth_fn = JuDGE.depth(tree) #define a function that returns the depth of each node in the tree
dpth = depth_fn(node) #get the depth of node in tree

JuDGE.history — Function

history(tree::AbstractTree)

Given tree, this function returns a function that takes a node of tree and returns that node's history back up to the root node of tree.

Required Arguments

tree is the tree that the history function will correspond to.

Example

history_fn = JuDGE.history(tree) #define a function that returns the history of each node in the tree

past = history_fn(node) #get a vector of nodes that precede node in tree

JuDGE.parent_builder — Function

parent_builder(tree::AbstractTree)

Given tree, this function returns a function that takes a node of tree and returns that node's parent.

Required Arguments

tree is the tree that the parent function will correspond to.

Example

parent_fn = JuDGE.parent_builder(tree) #define a function that returns the parent of each node in the tree
p = parent_fn(node) #get the parent of node

JuDGE Functions

JuDGE.JuDGEModel — Type

JuDGEModel(tree::AbstractTree,
           probabilities,
           sub_problem_builder::Function,
           solver;
           discount_factor::Float64,
           CVaR::Tuple{Float64,Float64},
           sideconstraints)

Define a JuDGE model.

Required arguments

tree is a reference to a scenario tree

probabilities is either a function, which returns a dictionary of the probabilities of all nodes in a tree, or simply the dictionary itself

sub_problem_builder is a function mapping a node to a JuMP model for each subproblems

solver is a reference to the optimizer used for the master problem (with appropriate settings); this can also be a tuple containing two optimizers (one for solving the relaxation, and one for solving the binary model)

Optional arguments

discount_factor is a number between 0 and 1 defining a constant discount factor along each arc in the scenario tree

CVaR is a tuple with the two CVaR parameters: (λ, β)

sideconstraints is a function which specifies side constraints in the master problem, see example * for further details.

Examples

judge = JuDGEModel(tree, ConditionallyUniformProbabilities, sub_problems,
                                Gurobi.Optimizer)
judge = JuDGEModel(tree, probabilities, sub_problems, CPLEX.Optimizer,
                                discount_factor=0.9, CVaR=(0.5,0.1)))

JuDGE.solve — Method

solve(judge::JuDGEModel;
      abstol = 10^-14,
      reltol = 10^-14,
      rlx_abstol = 10^-14,
      rlx_reltol = 10^-14,
      duration = Inf,
      iter = 2^63 - 1,
      inttol = 10^-9,
      allow_frac = 0,
      prune = Inf)

Solve a JuDGEModel judge without branch and price.

Required Arguments

judge is the JuDGE model that we wish to solve.

Optional Arguments

abstol is the absolute tolerance for the best integer-feasible objective value and the lower bound

reltol is the relative tolerance for the best integer-feasible objective value and the lower bound

rlx_abstol is the absolute tolerance for the relaxed master objective value and the lower bound

rlx_reltol is Set the relative tolerance for the relaxed master objective value and the lower bound

duration is the maximum duration

iter is the maximum number of iterations

inttol is the maximum deviation from 0 or 1 for integer feasible solutions

Used by the branch and price algorithm

allow_frac indicates wheither a fractional solution will be returned

prune is used to stop the algorithm before convergence, if a known upper bound for the problem is specified

Examples

JuDGE.solve(jmodel, rlx_abstol=10^-6)
JuDGE.solve(jmodel, abstol=10^-6)

JuDGE.branch_and_price — Function

branch_and_price(judge::JuDGEModel;
	  branch_method=JuDGE.constraint_branch,
	  search=:depth_first_resurface,
      abstol = 10^-14,
      reltol = 10^-14,
      rlx_abstol = 10^-14,
      rlx_reltol = 10^-14,
      duration = Inf,
      iter = 2^63 - 1,
      inttol = 10^-9)

Solve a JuDGEModel judge without branch and price.

Required Arguments

judge is the JuDGE model that we wish to solve.

Optional Arguments

branch_method specifies the way that constrants are added to create node nodes

search specifies the order in which nodes are solved in the (brnach-and-price) tree

abstol is the absolute tolerance for the best integer-feasible objective value and the lower bound

reltol is the relative tolerance for the best integer-feasible objective value and the lower bound

rlx_abstol is the absolute tolerance for the relaxed master objective value and the lower bound

rlx_reltol is Set the relative tolerance for the relaxed master objective value and the lower bound

duration is the maximum duration

iter is the maximum number of iterations

inttol is the maximum deviation from 0 or 1 for integer feasible solutions

Examples

JuDGE.branch_and_price(jmodel, abstol=10^-6)
JuDGE.branch_and_price(jmodel, branch_method=JuDGE.variable_branch, search=:lowestLB)

JuDGE.constraint_branch — Function

constraint_branch(master, tree, expansions, inttol)

This is an in-built function that is called during branch-and-price to perform a branch. Users can define their own functions that follow this format to create new branching strategies.

Required Arguments

master is the master problem of the JuDGE model

tree is the tree of the JuDGE model

expansions is a dictionary of capacity expansion variables, indexed by node.

inttol is the maximum permitted deviation from 0 or 1 for a value to still be considered binary.

JuDGE.resolve_subproblems — Function

resolve_subproblems(judge::JuDGEModel)

Once a JuDGE model has converged, it is necessary to re-solve the subproblems to find the optimal decisions within each node.

Required Arguments

jmodel is the JuDGE model that we wish to solve.

Examples

resolve_subproblems(judge)

JuDGE macros for subproblems

JuDGE.@expansion — Macro

expansion(model, variable, lag, span)

Defines an expansion variable variable within a subproblem model. Note that all subproblems must have the same set of expansion variables.

Required Arguments

model is the JuDGE subproblem that we are adding the expansion variable to

variable is the name of the variable being created, this will be automatically set to be binary; follows JuMP syntax if defining a set of variables.

Optional Arguments

lag is the number of nodes in the scenario between an expansion being decided, and it becoming available

span is the number of consecutive nodes in the scenario over which an expansion is available

Examples

@expansion(model, expand[1:5]) #defines an array of 5 variables with no lag, and unlimited lifespan
@expansion(model, expand[1:5,1:2], 1) #defines a matrix of 10 variables with a lag of 1, and unlimited lifespan
@expansion(model, expand, 0, 2) #defines a single variable with a lag of 0, and a lifespan of 2

JuDGE.@shutdown — Macro

shutdown(model, variable, lag, span)

Defines an shutdown variable variable within a subproblem model. Note that all subproblems must have the same set of shutdown variables.

Required Arguments

model is the JuDGE subproblem that we are adding the shutdown variable to

variable is the name of the variable being created, this will be automatically set to be binary; follows JuMP syntax if defining a set of variables.

Optional Arguments

lag is the number of nodes in the scenario between an shutdown being decided, and it becoming unavailable

span is the number of consecutive nodes in the scenario over which the shutdown will last

Examples

@shutdown(model, shut[1:5]) #defines an array of 5 variables with no lag, and unlimited duration
@shutdown(model, shut[1:5,1:2], 1) #defines a matrix of 10 variables with a lag of 1, and unlimited duration
@shutdown(model, shut, 0, 2) #defines a single variable with a lag of 0, and a lifespan of 2

JuDGE.@expansioncosts — Macro

expansioncosts(model, expr)

Defines a linear expression specifying the cost of expansions and shutdowns at the current node

Required Arguments

model is the JuDGE subproblem corresponding to the node in the scenario tree that we are adding specifying the costs for

expr is an AffExpr which gives the total cost of choosing expansion and shutdown variables at the current node

Example

@expansioncosts(model, sum(expand[i]*cost[node][i] for i in 1:5))

JuDGE.@maintenancecosts — Macro

maintenancecosts(model, expr)

Defines a linear expression specifying the ongoing cost of expansions and shutdowns available at the current node

Required Arguments

model is the JuDGE subproblem corresponding to the node in the scenario tree that we are adding specifying the costs for

expr is an AffExpr which gives the ongoing cost of expansions and shutdowns available at the current node

Example

@maintenancecosts(model, sum(expand[i]*ongoingcosts[node][i] for i in 1:5))

JuDGE.@sp_objective — Macro

sp_objective(model, expr)

Defines a linear expression specifying the cost of operating the system for the current node, excluding expansion or ongoing costs.

If it's possible to avoid costs by not using some previously expanded capacity, this can be included here with by directly including the expansion variable in the expression.

Required Arguments

model is the JuDGE subproblem corresponding to the node in the scenario tree that we are adding specifying the costs for

expr is an AffExpr which gives the subproblem costs.

Example

@sp_objective(model, sum(y[i]*c[node][i] for i in 1:5))

JuDGE Output

JuDGE.write_solution_to_file — Method

write_solution_to_file(jmodel::JuDGEModel,filename::String)

Given a JuDGE model and a filename, this function writes the entire solution to a CSV.

Required Arguments

jmodel is the JuDGE model whose solution we wish to write to a file

filename is the output filename

JuDGE.print_expansions — Method

print_expansions(jmodel::JuDGEModel;
                onlynonzero::Bool=true,
                inttol=10^-9,
                format=nothing)

Given a solved JuDGE model, this function will write the optimal capacity expansion decisions to the REPL.

Required Arguments

jmodel is the JuDGE model whose solution we wish to write to a file

Optional Arguments

onlynonzero is a boolean, if set to true the function will only print expansions with a non-zero value.

inttol is the integrality tolerance; any expansion variable value less than this will be treated as 0, and any value greater than 1-inttol will be treated as 1

format is a function that specifies customised printing of expansion values. See the examples for more details.

Deterministic Equivalent

JuDGE.DetEqModel — Type

DetEqModel(tree::AbstractTree,
           probabilities,
           sub_problem_builder::Function,
           solver;
           discount_factor::Float64,
           CVaR::Tuple{Float64,Float64},
           sideconstraints)

Define a deterministic equivalent model for the stochastic capacity expansion problem.

Required arguments

tree is a reference to a scenario tree

probabilities is either a function, which returns a dictionary of the probabilities of all nodes in a tree, or simply the dictionary itself

sub_problem_builder is a function mapping a node to a JuMP model for each subproblems

solver is a reference to the optimizer used for this problem (with appropriate settings)

Optional arguments

discount_factor is a number between 0 and 1 defining a constant discount factor along each arc in the scenario tree

CVaR is a tuple with the two CVaR parameters: (λ, β)

sideconstraints is a function which specifies side constraints in the master problem, see example * for further details.

Examples

deteq = DetEqModel(tree, ConditionallyUniformProbabilities, sub_problems,
                                Gurobi.Optimizer)
judge = DetEqModel(tree, probabilities, sub_problems, CPLEX.Optimizer,
                                discount_factor=0.9, CVaR=(0.5,0.1)))

JuDGE.solve — Method

solve(deteq::DetEqModel)

Solve a determinisitc equivalent model.

Required Arguments

deteq is the determinisitc equivalent model that we wish to solve.

Example

JuDGE.solve(deteq)

Deterministic Equivalent Output

JuDGE.write_solution_to_file — Method

write_solution_to_file(deteq::DetEqModel,filename::String)

Given a deterministic equivalent model and a filename, this function writes the entire solution to a CSV.

Required Arguments

deteq is the deterministic equivalent model whose solution we wish to write to a file

filename is the output filename

JuDGE.print_expansions — Method

print_expansions(deteq::DetEqModel;
                onlynonzero::Bool=true,
                inttol=10^-9,
                format=nothing)

Given a solved deterministic equivalent model, this function will write the optimal capacity expansion decisions to the REPL.

Required Arguments

deteq is the deterministic equivalent model whose solution we wish to write to a file

Optional Arguments

onlynonzero is a boolean, if set to true the function will only print expansions with a non-zero value.

inttol is the integrality tolerance; any expansion variable value less than this will be treated as 0, and any value greater than 1-inttol will be treated as 1

format is a function that specifies customised printing of expansion values. See the examples for more details.