AbstractTree API

Defining Trees

JuDGE.narytree — Function

narytree(depth::Int64, degree::Int64)

Given the depth and degree, this function returns an N-ary tree. Note that a depth of 0 return a single Leaf node (which is also the root node of the tree).

Required Arguments

depth is the maximum number of arcs from the root node any Leaf node

degree is the number of children of all nodes, other than the Leaf nodes

Example

tree = narytree(2,2)

JuDGE.tree_from_file — Function

tree_from_file(filename::String)

Construct tree from a file, each line in the file is of the form B,A,... representing an arc in the tree, from node "A" to node "B". The total number of columns is arbitrary. The first row of the file should be n,p,... these column headers are converted into symbols used to index the data. Each column itself converted into a dictionary, indexed by the node.

Required Arguments

string is the full path of the file containing the tree

Example

tree, data = tree_from_file(joinpath(@__DIR__,"tree.csv"))

JuDGE.tree_from_leaves — Function

tree_from_leaves(leafnodes::Array{Array{Int64,1},1}, probs::Array{Float64,1})

Construct tree from Array of leaf nodes, and (optionally) the corresponding probabilities

Required Arguments

leafnodes is an array of arrays defining the set of leaf nodes

Optional Arguments

probs is an array of probabilities for the leaf nodes

Example

(tree,prob) = tree_from_leaves([[1,1,1],[1,1,2],[1,2,1],[1,2,2]],[0.25,0.25,0.25,0.25])
tree = tree_from_leaves([[1,1,1],[1,1,2],[1,2,1],[1,2,2]])

JuDGE.print_tree — Function

print_tree(some_tree::AbstractTree, data::Dict{AbstractTree,Float64})

Given some_tree, this function prints a simple representation of the tree to the REPL.

Required Arguments

some_tree is the tree we wish to visualise

Optional Arguments

data is a dictionary indexed by the nodes of some_tree

Nodes of Trees

Base.collect — Function

collect(tree::Tree;order=:depth)

Given tree, this function returns an array of corresponding nodes. By default this will be in a depth-first order.

Required Arguments

tree is the tree from which we wish to collect the nodes

Optional Arguments

order can be set to :depth or :breadth to specify the order that the nodes are listed in the array.

Examples

nodes = collect(tree) #gets an array of nodes from tree in depth-first order
nodes = collect(tree,order=:breadth) #gets an array of nodes from tree in breadth-first order

JuDGE.get_leafnodes — Function

get_leafnodes(tree::AbstractTree)

Given tree, this function returns an array of corresponding Leaf nodes.

Required Arguments

tree is the tree from which we wish to collect leaf nodes

Example

leafnodes = JuDGE.get_leafnodes(tree) #define a function that returns the parent of each node in the tree

JuDGE.get_node — Function

get_node(tree::AbstractTree, indices::Array{Int64,1})

Given a tree, and an array of indices, this function returns the corresponding node in the tree.

Required Arguments

tree is the tree from which we are finding the node

indicies is an array of integer indices identifying a node within tree.

Examples

node = get_node(tree,[1]) #get the root node
node = get_node(tree,[1,1]) #get the first child of the root node
node = get_node(tree,[1,2]) #get the second child of the root node

Tree Probabilities

JuDGE.convert_probabilities — Function

convert_probabilities(tree::AbstractTree, probabilities::Dict{AbstractTree,Float64})

Given a dictionary of conditional probabilities for each node in tree, this function returns a dictionary that maps each node of tree to the corresponding unconditional probability.

Required Arguments

tree is the tree that the probabilities pertain to

probabilities is a dictionary of condition probabilities for each node in tree

Example

probs = JuDGE.convert_probabilities(tree,probabilities)

JuDGE.ConditionallyUniformProbabilities — Function

ConditionallyUniformProbabilities(tree::AbstractTree)

Given a tree, this function returns a dictionary which maps nodes of the tree to probabilities, given that there are conditionally uniform probabilities over the children of any node.

Required Arguments

tree is the tree for which the probability distribution will be generated

Example

probs = ConditionallyUniformProbabilities(tree)

JuDGE.UniformLeafProbabilities — Function

UniformLeafProbabilities(tree::AbstractTree)

Given a tree, this function returns a dictionary which maps nodes of the tree to probabilities, given that there is a uniform distribution over the leaf nodes

Required Arguments

tree is the tree for which the probability distribution will be generated

Example

probs = UniformLeafProbabilities(tree)

Other Tree functions

JuDGE.depth — Function

depth(tree::AbstractTree)

Given tree, this function returns a function that takes a subtree of tree and returns that subtree's history back up to the root node of tree.

Required Arguments

tree is the tree that the history function will correspond to.

Example

depth_fn = JuDGE.history(tree) #define a function that returns the depth of each node in the tree
dpth = depth_fn(node) #get the depth of node in tree

JuDGE.history — Function

history(tree::AbstractTree)

Given tree, this function returns a function that takes a node of tree and returns that node's history back up to the root node of tree.

Required Arguments

tree is the tree that the history function will correspond to.

Example

history_fn = JuDGE.history(tree) #define a function that returns the history of each node in the tree

past = history_fn(node) #get a vector of nodes that precede node in tree

JuDGE.parent_builder — Function

parent_builder(tree::AbstractTree)

Given tree, this function returns a function that takes a node of tree and returns that node's parent.

Required Arguments

tree is the tree that the parent function will correspond to.

Example

parent_fn = JuDGE.parent_builder(tree) #define a function that returns the parent of each node in the tree
p = parent_fn(node) #get the parent of node