JuDGE API

JuDGE Functions

JuDGE.JuDGEModel — Type

JuDGEModel(tree::AbstractTree,
           probabilities,
           sub_problem_builder::Function,
           solver;
           discount_factor::Float64,
           CVaR::Tuple{Float64,Float64},
           sideconstraints)

Define a JuDGE model.

Required arguments

tree is a reference to a scenario tree

probabilities is either a function, which returns a dictionary of the probabilities of all nodes in a tree, or simply the dictionary itself

sub_problem_builder is a function mapping a node to a JuMP model for each subproblems

solver is a reference to the optimizer used for the master problem (with appropriate settings); this can also be a tuple containing two optimizers (one for solving the relaxation, and one for solving the binary model)

Optional arguments

discount_factor is a number between 0 and 1 defining a constant discount factor along each arc in the scenario tree

CVaR is a tuple with the two CVaR parameters: (λ, β)

sideconstraints is a function which specifies side constraints in the master problem, see example * for further details.

Examples

judge = JuDGEModel(tree, ConditionallyUniformProbabilities, sub_problems,
                                Gurobi.Optimizer)
judge = JuDGEModel(tree, probabilities, sub_problems, CPLEX.Optimizer,
                                discount_factor=0.9, CVaR=(0.5,0.1)))

JuDGE.solve — Method

solve(judge::JuDGEModel;
      abstol = 10^-14,
      reltol = 10^-14,
      rlx_abstol = 10^-14,
      rlx_reltol = 10^-14,
      duration = Inf,
      iter = 2^63 - 1,
      inttol = 10^-9,
      allow_frac = 0,
      prune = Inf)

Solve a JuDGEModel judge without branch and price.

Required Arguments

judge is the JuDGE model that we wish to solve.

Optional Arguments

abstol is the absolute tolerance for the best integer-feasible objective value and the lower bound

reltol is the relative tolerance for the best integer-feasible objective value and the lower bound

rlx_abstol is the absolute tolerance for the relaxed master objective value and the lower bound

rlx_reltol is Set the relative tolerance for the relaxed master objective value and the lower bound

duration is the maximum duration

iter is the maximum number of iterations

inttol is the maximum deviation from 0 or 1 for integer feasible solutions

Used by the branch and price algorithm

allow_frac indicates wheither a fractional solution will be returned

prune is used to stop the algorithm before convergence, if a known upper bound for the problem is specified

Examples

JuDGE.solve(jmodel, rlx_abstol=10^-6)
JuDGE.solve(jmodel, abstol=10^-6)

JuDGE.resolve_subproblems — Function

resolve_subproblems(judge::JuDGEModel)

Once a JuDGE model has converged, it is necessary to re-solve the subproblems to find the optimal decisions within each node.

Required Arguments

jmodel is the JuDGE model that we wish to solve.

Examples

resolve_subproblems(judge)

JuDGE macros for subproblems

JuDGE.@expansion — Macro

expansion(model, variable, lag, span)

Defines an expansion variable variable within a subproblem model. Note that all subproblems must have the same set of expansion variables.

Required Arguments

model is the JuDGE subproblem that we are adding the expansion variable to

variable is the name of the variable being created, this will be automatically set to be binary; follows JuMP syntax if defining a set of variables.

Optional Arguments

lag is the number of nodes in the scenario between an expansion being decided, and it becoming available

span is the number of consecutive nodes in the scenario over which an expansion is available

Examples

@expansion(model, expand[1:5]) #defines an array of 5 variables with no lag, and unlimited lifespan
@expansion(model, expand[1:5,1:2], 1) #defines a matrix of 10 variables with a lag of 1, and unlimited lifespan
@expansion(model, expand, 0, 2) #defines a single variable with a lag of 0, and a lifespan of 2

JuDGE.@shutdown — Macro

shutdown(model, variable, lag, span)

Defines an shutdown variable variable within a subproblem model. Note that all subproblems must have the same set of shutdown variables.

Required Arguments

model is the JuDGE subproblem that we are adding the shutdown variable to

variable is the name of the variable being created, this will be automatically set to be binary; follows JuMP syntax if defining a set of variables.

Optional Arguments

lag is the number of nodes in the scenario between an shutdown being decided, and it becoming unavailable

span is the number of consecutive nodes in the scenario over which the shutdown will last

Examples

@shutdown(model, shut[1:5]) #defines an array of 5 variables with no lag, and unlimited duration
@shutdown(model, shut[1:5,1:2], 1) #defines a matrix of 10 variables with a lag of 1, and unlimited duration
@shutdown(model, shut, 0, 2) #defines a single variable with a lag of 0, and a lifespan of 2

JuDGE.@expansioncosts — Macro

expansioncosts(model, expr)

Defines a linear expression specifying the cost of expansions and shutdowns at the current node

Required Arguments

model is the JuDGE subproblem corresponding to the node in the scenario tree that we are adding specifying the costs for

expr is an AffExpr which gives the total cost of choosing expansion and shutdown variables at the current node

Example

@expansioncosts(model, sum(expand[i]*cost[node][i] for i in 1:5))

JuDGE.@maintenancecosts — Macro

maintenancecosts(model, expr)

Defines a linear expression specifying the ongoing cost of expansions and shutdowns available at the current node

Required Arguments

model is the JuDGE subproblem corresponding to the node in the scenario tree that we are adding specifying the costs for

expr is an AffExpr which gives the ongoing cost of expansions and shutdowns available at the current node

Example

@maintenancecosts(model, sum(expand[i]*ongoingcosts[node][i] for i in 1:5))

JuDGE.@sp_objective — Macro

sp_objective(model, expr)

Defines a linear expression specifying the cost of operating the system for the current node, excluding expansion or ongoing costs.

If it's possible to avoid costs by not using some previously expanded capacity, this can be included here with by directly including the expansion variable in the expression.

Required Arguments

model is the JuDGE subproblem corresponding to the node in the scenario tree that we are adding specifying the costs for

expr is an AffExpr which gives the subproblem costs.

Example

@sp_objective(model, sum(y[i]*c[node][i] for i in 1:5))

Deterministic Equivalent

JuDGE.DetEqModel — Type

DetEqModel(tree::AbstractTree,
           probabilities,
           sub_problem_builder::Function,
           solver;
           discount_factor::Float64,
           CVaR::Tuple{Float64,Float64},
           sideconstraints)

Define a deterministic equivalent model for the stochastic capacity expansion problem.

Required arguments

tree is a reference to a scenario tree

probabilities is either a function, which returns a dictionary of the probabilities of all nodes in a tree, or simply the dictionary itself

sub_problem_builder is a function mapping a node to a JuMP model for each subproblems

solver is a reference to the optimizer used for this problem (with appropriate settings)

Optional arguments

discount_factor is a number between 0 and 1 defining a constant discount factor along each arc in the scenario tree

CVaR is a tuple with the two CVaR parameters: (λ, β)

sideconstraints is a function which specifies side constraints in the master problem, see example * for further details.

Examples

deteq = DetEqModel(tree, ConditionallyUniformProbabilities, sub_problems,
                                Gurobi.Optimizer)
judge = DetEqModel(tree, probabilities, sub_problems, CPLEX.Optimizer,
                                discount_factor=0.9, CVaR=(0.5,0.1)))

JuDGE.solve — Method

solve(deteq::DetEqModel)

Solve a determinisitc equivalent model.

Required Arguments

deteq is the determinisitc equivalent model that we wish to solve.

Example

JuDGE.solve(deteq)

Solution Output Functions

JuDGE Output

JuDGE.write_solution_to_file — Method

write_solution_to_file(jmodel::JuDGEModel,filename::String)

Given a JuDGE model and a filename, this function writes the entire solution to a CSV.

Required Arguments

jmodel is the JuDGE model whose solution we wish to write to a file

filename is the output filename

JuDGE.print_expansions — Method

print_expansions(jmodel::JuDGEModel;
                onlynonzero::Bool=true,
                inttol=10^-9,
                format=nothing)

Given a solved JuDGE model, this function will write the optimal capacity expansion decisions to the REPL.

Required Arguments

jmodel is the JuDGE model whose solution we wish to write to a file

Optional Arguments

onlynonzero is a boolean, if set to true the function will only print expansions with a non-zero value.

inttol is the integrality tolerance; any expansion variable value less than this will be treated as 0, and any value greater than 1-inttol will be treated as 1

format is a function that specifies customised printing of expansion values. See the examples for more details.

Deterministic Equivalent Output

JuDGE.write_solution_to_file — Method

write_solution_to_file(deteq::DetEqModel,filename::String)

Given a deterministic equivalent model and a filename, this function writes the entire solution to a CSV.

Required Arguments

deteq is the deterministic equivalent model whose solution we wish to write to a file

filename is the output filename

JuDGE.print_expansions — Method

print_expansions(deteq::DetEqModel;
                onlynonzero::Bool=true,
                inttol=10^-9,
                format=nothing)

Given a solved deterministic equivalent model, this function will write the optimal capacity expansion decisions to the REPL.

Required Arguments

deteq is the deterministic equivalent model whose solution we wish to write to a file

Optional Arguments

onlynonzero is a boolean, if set to true the function will only print expansions with a non-zero value.

inttol is the integrality tolerance; any expansion variable value less than this will be treated as 0, and any value greater than 1-inttol will be treated as 1

format is a function that specifies customised printing of expansion values. See the examples for more details.